change the rotation, add the noise, train h1
