+++ sim/resources/h1_2/joints.py
@@ -96,10 +96,10 @@ class Legs(Node):
 class Robot(Node):
     height = 0.9
     rotation = [
+        0,
+        0,
+        0,
         1,
-        0,
-        0,
-        0,
     ]
 
     torso = Torso()
