+++ sim/utils/math.py
@@ -32,7 +32,7 @@
 
 import numpy as np
 
-from isaacgym.torch_utils import quat_apply, normalize  # isort: skip
+from isaacgym.torch_utils import quat_apply, normalize, get_euler_xyz  # isort: skip
 import torch  # isort: skip
 
 
@@ -58,3 +58,11 @@ def torch_rand_sqrt_float(lower, upper, shape, device):
     r = torch.where(r < 0.0, -torch.sqrt(-r), torch.sqrt(r))
     r = (r + 1.0) / 2.0
     return (upper - lower) * r + lower
+
+
+def get_euler_xyz_tensor(quat):
+    r, p, w = get_euler_xyz(quat)
+    # stack r, p, w in dim1
+    euler_xyz = torch.stack((r, p, w), dim=1)
+    euler_xyz[euler_xyz > np.pi] -= 2 * np.pi
+    return euler_xyz
